Updated security of queries

+++ products/views.py
@@ -493,20 +493,22 @@ def add_query(request):
 
 def accept_query(request, Query_id):
     query = ExtraItem.objects.get(id=Query_id)
-    query.flag = "Accepted"
-    query.save()
+    if query.customer == request.user:
+        query.flag = "Accepted"
+        query.save()
     return HttpResponseRedirect(reverse('shop-details', kwargs={'product_id': query.product.id}))
-
 
 def decline_query(request, Query_id):
     query = ExtraItem.objects.get(id=Query_id)
-    query.flag = "Rejected"
-    query.save()
+    if query.customer == request.user:
+        query.flag = "Rejected"
+        query.save()
     return HttpResponseRedirect(reverse('shop-details', kwargs={'product_id': query.product.id}))
 
 def delete_query(request, Query_id):
     query = ExtraItem.objects.get(id=Query_id)
-    query.delete()
+    if query.customer == request.user:
+        query.delete()
     return HttpResponseRedirect(reverse('shop-details', kwargs={'product_id': query.product.id}))
 
 def notifications_view(request):
